// server/src/sources/devto/devto.types.ts
export type DevtoRaw = {
  id: number;
  title: string;
  description: string;
  url: string;
  comments_count: number;
  public_reactions_count: number;
  published_timestamp: "2026-01-22T14:05:49Z";
  positive_reactions_count: number;
  cover_image: string;
  reading_time_minutes: number;
  tag_list: string[];
  user: {
    name: string;
    username: string;
    twitter_username: string;
    github_username: string;
  };
};